# public/docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Call Guru API",
        "description": "Authentication APIs for mobile OTP flow.",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api",
            "description": "API Base URL"
        }
    ],
    "paths": {
        "/v1/auth/send-otp": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Send OTP",
                "operationId": "007aceb0e800b3e130b0a7b80764a5a9",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "countryCode",
                                    "mobileNumber"
                                ],
                                "properties": {
                                    "countryCode": {
                                        "type": "string",
                                        "example": "+91"
                                    },
                                    "mobileNumber": {
                                        "type": "string",
                                        "example": "9876543210"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OTP sent",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "OTP sent successfully"
                                        },
                                        "otp": {
                                            "type": "string",
                                            "example": "123456"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Invalid mobile",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Invalid mobile number"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/v1/auth/verify-otp": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Verify OTP",
                "operationId": "13daa983c70d922bb2070233aff09593",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "countryCode",
                                    "mobileNumber",
                                    "otp"
                                ],
                                "properties": {
                                    "countryCode": {
                                        "type": "string",
                                        "example": "+91"
                                    },
                                    "mobileNumber": {
                                        "type": "string",
                                        "example": "9876543210"
                                    },
                                    "otp": {
                                        "type": "string",
                                        "example": "123456"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OTP verified",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "oneOf": [
                                        {
                                            "properties": {
                                                "success": {
                                                    "type": "boolean",
                                                    "example": true
                                                },
                                                "message": {
                                                    "type": "string",
                                                    "example": "Success"
                                                },
                                                "id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "name": {
                                                    "type": "string",
                                                    "example": "Rahul"
                                                },
                                                "accessToken": {
                                                    "type": "string",
                                                    "example": "1|sanctum_access_token_here"
                                                },
                                                "refreshToken": {
                                                    "type": "string",
                                                    "example": "2|sanctum_refresh_token_here"
                                                }
                                            },
                                            "type": "object"
                                        },
                                        {
                                            "properties": {
                                                "success": {
                                                    "type": "boolean",
                                                    "example": true
                                                },
                                                "message": {
                                                    "type": "string",
                                                    "example": "User registration required"
                                                },
                                                "isNewUser": {
                                                    "type": "boolean",
                                                    "example": true
                                                }
                                            },
                                            "type": "object"
                                        }
                                    ]
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Invalid or expired OTP",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Invalid OTP"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Inactive user account",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Your account is inactive. Please contact the administrator."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/v1/auth/register": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Register user",
                "operationId": "501189c078fa4e0494985e8e952b7f61",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "countryCode",
                                    "mobileNumber",
                                    "name"
                                ],
                                "properties": {
                                    "countryCode": {
                                        "type": "string",
                                        "example": "+91"
                                    },
                                    "mobileNumber": {
                                        "type": "string",
                                        "example": "9876543210"
                                    },
                                    "name": {
                                        "type": "string",
                                        "example": "Rahul"
                                    },
                                    "date_of_birth": {
                                        "type": "string",
                                        "format": "date-time",
                                        "example": "1996-05-28T00:00:00.000Z"
                                    },
                                    "sex": {
                                        "type": "string",
                                        "example": "Male",
                                        "enum": [
                                            "Male",
                                            "Female",
                                            "Other"
                                        ]
                                    },
                                    "language": {
                                        "type": "integer",
                                        "example": 1,
                                        "nullable": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Registered",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Success"
                                        },
                                        "id": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "name": {
                                            "type": "string",
                                            "example": "Test User"
                                        },
                                        "accessToken": {
                                            "type": "string",
                                            "example": "7|sanctum_access_token_here"
                                        },
                                        "refreshToken": {
                                            "type": "string",
                                            "example": "8|sanctum_refresh_token_here"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "409": {
                        "description": "Already exists",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "User already exists"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/v1/auth/refresh-token": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Refresh access token",
                "operationId": "7c5de428964f52b67149203c22b3dbb0",
                "responses": {
                    "200": {
                        "description": "Token refreshed",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Success"
                                        },
                                        "accessToken": {
                                            "type": "string",
                                            "example": "9|sanctum_access_token_here"
                                        },
                                        "refreshToken": {
                                            "type": "string",
                                            "example": "10|sanctum_refresh_token_here"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Invalid refresh token or inactive account",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Invalid refresh token."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/v1/auth/logout": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Logout",
                "operationId": "e4e9f7c8b49c424ac3b9afb2c0ff6815",
                "responses": {
                    "200": {
                        "description": "Logged out",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Logged out successfully"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        }
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "description": "Use Bearer token for authenticated endpoints",
                "scheme": "bearer",
                "bearerFormat": "Token"
            }
        }
    },
    "tags": [
        {
            "name": "Auth",
            "description": "OTP based authentication endpoints"
        }
    ]
}